+++ packages/extensions/src/list/index.ts
@@ -1,33 +1,21 @@
 import {
-  defineCommands,
   defineNodeSpec,
   definePlugin,
-  insertNode,
   union,
   type Extension,
 } from '@prosekit/core'
-import type { Command } from '@prosekit/pm/state'
 import {
   createListPlugins,
   createListSpec,
-  createDedentListCommand as dedentList,
-  createIndentListCommand as indentList,
   listInputRules,
-  createMoveListCommand as moveList,
-  createSplitListCommand as splitList,
-  createToggleCollapsedCommand as toggleCollapsed,
-  createToggleListCommand as toggleList,
-  createUnwrapListCommand as unwrapList,
-  createWrapInListCommand as wrapInList,
-  type ListAttributes,
 } from 'prosemirror-flat-list'
 
 import { defineInputRule } from '../input-rule'
 
+import { defineListCommands } from './commands'
 import { defineListKeymap } from './keymap'
 
 export { ListDOMSerializer } from 'prosemirror-flat-list'
-
 export type {
   DedentListOptions,
   IndentListOptions,
@@ -39,27 +27,6 @@ export type {
 
 export function defineListSpec() {
   return defineNodeSpec({ ...createListSpec(), name: 'list' })
-}
-
-function insertList(attrs?: ListAttributes): Command {
-  return insertNode({ type: 'list', attrs })
-}
-
-/**
- * Define list commands
- */
-export function defineListCommands() {
-  return defineCommands({
-    dedentList,
-    indentList,
-    moveList,
-    splitList,
-    toggleCollapsed,
-    toggleList,
-    unwrapList,
-    wrapInList,
-    insertList,
-  })
 }
 
 export function defineListPlugins() {
@@ -83,4 +50,4 @@ export function defineList() {
   ])
 }
 
-export { defineListKeymap }
+export { defineListCommands, defineListKeymap }
